requirements.txt
flask
openai
python-dotenv
duckduckgo-search
